created node scanning function (node_scan) in tmemscan.c
