+++ src/tmemscan.h
@@ -1,0 +1,9 @@
+/* mnscan header.h */
+#ifndef TMEMSCAN_HEADER_H
+#define TMEMSCAN_HEADER_H
+
+int avail_nodes(void);
+
+int avail_pages(void);
+
+#endif /* TMEMSCAN_HEADER_H */
