--- src/vmmodscan.h
@@ -1,9 +1,0 @@
-/* mnscan header.h */
-#ifndef VMMODSCAN_HEADER_H
-#define VMMODSCAN_HEADER_H
-
-int avail_nodes(void);
-
-int avail_pages(void);
-
-#endif /* VMMODSCAN_HEADER_H */
